Rename legacy macros fallback rule to keep convention
