+++ chimney/src/main/scala-2/io/scalaland/chimney/internal/compiletime/derivation/transformer/DerivationPlatform.scala
@@ -5,8 +5,8 @@ import io.scalaland.chimney.internal.compiletime.DefinitionsPlatform
 private[derivation] trait DerivationPlatform
     extends Derivation
     with rules.TransformSubtypesRuleModule
-    with DerivationWithLegacyMacros {
+    with rules.LegacyMacrosFallbackRuleModule {
   this: DefinitionsPlatform =>
 
-  override protected val rulesAvailableForPlatform: Seq[Rule] = Seq(TransformSubtypesRule, LegacyMacrosRule)
+  override protected val rulesAvailableForPlatform: Seq[Rule] = Seq(TransformSubtypesRule, LegacyMacrosFallbackRule)
 }
